Create scrobble tracks job
This is the most disgusting mess ever, please don't judge me. The
`Track` objects in RSpotify apparently can't be compared with == or
maybe I am dumb and using it wrong, but it means I have to compare
myself. I just compared song name and artist which seems fine to me, but
maybe I should just check the ID. I want to avoid double scrobbling if
there is the same song as a single and on the album so screw it.