Replace current spotify user with helper
This will be user literally everywhere, so lets just take it into its
own method. It also makes sense to raise in this case since there will
really never be a case where we want a spotify user and if we can't find
one we still want to do stuff. That is a hard failure.