Refactor scrobbling
So the old way ths was done was gross and messy. This is also gross and
messy and is also totally unscalable.... but I think it is more
reliable. Since I check every minute and only check the last song it is
way less likely I miss a song, but running a job for every user every
single minute is... not great.