Make scrobble track not retry and not fail
So as far as I can tell this garbage doesn't work anyways but I don't
have the time to make it work right now and this seems to help a tiny
bit, so here we are. I don't think it picks up the retry: false at all,
but I will need to figure out why. Also, I know I shouldn't rescue
standard error, but I just want this to not retry constantly on top of
already running way too often for some reason. Sue me.