Add artist inputs
This adds boxes for users to input their own artists. These artists are
then searched on spotify and used for recommendations. As of right now
if no artist is found based on the searched string, it just ignores that
and gives the user an error message so they can hopefully sort it out.