Fix user login issues
Turns out that job was creating tons of spotify users for the one user,
which is really stupid by me. It should never run again and it is not
something that can be done anyways, so I am killing the code. Now when a
user auths, it will check if they have an existing spotify user, and if
so just override that one instead of making a new one and confusing
things.