Add chart to overview page
This is super bare-bones but I need something on this page. This leaves
a bunch of comments and garbage in here, but I just need something up
there for now. The intention is to add more charts and allow people to
change whether they want to see the last month/year/week, etc.