Fix my migration
Ok, this does the thing you aren't supposed to do, modify an existing
migration. This is me just testing how to make this shit actually work,
now that it does, lets see if we can do this in place since the data is
already there and I don't want to move it all over again.