Add top 5 artists of last 6 months
This needs a name rewrite for sure, but this puts in the basics of a
chart I want to implement. It is sort of impossible to test manually at
the moment since my DB is basically empty, but I _think_ it works, and
when I use it for a chart for real I will make sure I actually have data
to try it with.