Update home page
The photo was kind of a joke since it changes a lot and I need a user to
login every time to get the updated on, so screw that. Also, this page
was just ugly as hell. This makes it slightly better, and also uses DB
data instead of hitting the network if we can.