Rebuild the home page from scratch
This looked like trash, so kendra took over and gave it a nice touch up.
This shows some basic data, but if a user has no data it is super ugly.
The hope is that we will put up a PR next to say if you have barely any
listens, either keep listening to music or upload your data through the
upload page.